refactor: rename spark-backup-format to backup-payload
